pyproject.toml
[project]
name = "streamlit-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.3",
    "google-generativeai>=0.8.4",
    "langchain>=0.3.19",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.18",
    "langchain-google-genai>=2.0.9",
    "langchain-mcp-adapters>=0.1.4",
    "markdown>=3.7",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.4.1",
    "sentencepiece>=0.2.0",
    "streamlit>=1.42.2",
    "streamlit-authenticator>=0.4.1",
    "unstructured>=0.16.23",
]
